django/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-solo==2.0.0
idna==3.4
psycopg2==2.9.6
requests==2.29.0
sqlparse==0.4.4
urllib3==1.26.15
